Introduce StopAndReset strategy
Fix issues with failed processes after /unload and processes that
need to be removed with SIGKILL. When a Process is SIGKILL'd it will
wind up in a StateFailed state due to complex interactions between
various contexts and channels for managing state.

The StopAndReset strategy will call Process.StopImmediately() to
terminate the process and then replace it with NewProcess().