proxy: fix race in TestProxyManager_StreamingEndpointsReturnNoBufferingHeader
This race condition only appeared in the test and not part of the
production code. The test was running the HTTP handler in a goroutine
and then immediately reading from httptest.ResponseRecorder while the
handler was still potentially writing to it. The fix was to remove the
time.Sleep and use channels to signal progress